# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: ClawBench API
  description: |
    ClawBench 的完整 HTTP API。所有 `/api/` 路由默认经 `middleware.Auth`
    鉴权（会话 Cookie，或本机 AI 子进程的短时 `aiToken`），少数端点刻意免鉴权
    （见各 operation 的 `security: []`）。
    聊天流式内容通过 WebSocket 推送（见 `/api/ai/events/ws`），不在本文件建模。

    约定：
    - 统一错误体见 `components/schemas/ErrorResponse`。
    - 项目范围通过 `clawbench_project` Cookie 或 `project_path` 查询参数传递。
    - 时间戳为 epoch 毫秒（除显式说明外）。
  version: "1.0.0"
servers:
  - url: /
    description: 同源（文件管理器内的 Swagger UI 直接可用）
...
security:
  - cookieAuth: []
  - aiToken: []
paths:
...
  /api/rag/search:
    post:
      tags: [RAG]
      summary: 语义 / 全文混合搜索
      description: |
        搜索历史对话（排除 thinking 与 tool_use 块）。
        项目范围由项目 Cookie 决定：Cookie 值为任意项目路径时只搜该项目，
        不发送 Cookie 时跨全部项目（仅本机 AI 令牌可省略 Cookie）。
      operationId: ragSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                q:
                  type: string
                  description: 搜索词（必填）。
                limit:
                  type: integer
                backend:
                  type: string
                role:
                  type: string
                session_id:
                  type: string
                exclude_session_id:
                  type: string
                from:
                  type: string
                to:
                  type: string
      responses:
        "200":
          description: 匹配分块
...
  /api/rag/session-search:
    post:
      tags: [RAG]
      summary: 会话聚合搜索
      operationId: ragSessionSearch
      description: |
        跨会话聚合搜索并返回命中位置，供会话搜索抽屉高亮。
        `q` 为空时进入「浏览全部」模式：列出该项目下的会话，此时
        `archived`/`session_type`/`sort`/`cursor` 仍然生效。
        浏览态不会混排两类会话：`all`/`chat` 均列出对话会话，只有显式
        `task` 才切换到任务执行会话。
        项目范围由项目 Cookie 决定：Cookie 值为任意项目路径时只搜该项目，
        不发送 Cookie 时跨全部项目（仅本机 AI 令牌可省略 Cookie）。
        同时匹配会话名称与消息内容，名称命中排在内容命中之前；两通道都命中的
        会话只出现一次，且共用同一组过滤。按 `newest`/`oldest` 排序时名称置顶
        不再保留。RAG 未配置时仍可用名称匹配，此时 `mode` 为 `title`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                q:
                  type: string
                  description: 搜索词；为空时列出会话而非搜索。
                backend:
                  type: string
                role:
                  type: string
                session_id:
                  type: string
                exclude_session_id:
                  type: string
                from:
                  type: string
                to:
                  type: string
                prefer_mode:
                  type: string
                  description: 强制检索模式，如 `fts`；缺省为混合检索。
                archived:
                  type: string
                  enum: [all, active, archived]
                  description: 归档筛选（字符串，非布尔）。默认 `all`。
                session_type:
                  type: string
                  enum: [all, chat, task]
                  description: |
                    会话类型筛选。`task` 对应库中的 `scheduled` 会话（计划任务的
                    单次执行会话），`chat` 为交互式对话。默认 `all`，即不做类型
                    限制（搜索态下两类都会返回）。
                    浏览态（`q` 为空）一次只列出一类：`all`/`chat` 列出对话，
                    `task` 列出任务执行。
                sort:
                  type: string
                  enum: [newest, oldest, relevance]
                  description: 排序方式，默认 `relevance`（未知值也回落到它）。
                cursor:
                  type: string
                  description: 游标时间戳；需与 `cursor_id` 同时提供才生效。
                cursor_id:
                  type: string
      responses:
        "200":
          description: |
            聚合结果。会话带 `title_match`、`title_match_positions`（标题内
            字符级偏移，用于高亮）与 `title_only`（无内容分块，需懒加载首条消息）。
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: clawbench_session
      description: 登录后下发的会话 Cookie；CLI 使用 `<DataDir>/cookie-token`。
    aiToken:
      type: apiKey
      in: header
      name: X-ClawBench-AI-Token
      description: |
        本机 AI 子进程使用的短时令牌，由内置斜杠命令（`/cb-task` 等）在注入提示词时签发，
        有效期 30 分钟，以 `CookieToken` 做 HMAC 签名。仅接受来自回环地址的请求：
        地址与签名须同时成立。地址单独不再放行——FRP 隧道的 frpc 亦从 `127.0.0.1` 回拨。
...